create image component separate from books page
